Move StrategoEvaluationStategy to core
